phandle != int in prom.c
GCC this pointed out:
second/prom.c: In function ‘prom_get_netinfo’:
second/prom.c:692: warning: format ‘%d’ expects type ‘int’, but argument 2 has type ‘prom_handle’

+++ second/prom.c
@@ -715,7 +715,7 @@ struct bootp_packet * prom_get_netinfo (void)
 
      chosen = prom_finddevice("/chosen");
      if (chosen < 0) {
-          DEBUG_F("chosen=%d\n", chosen);
+          DEBUG_F("chosen=%lu\n", (unsigned long)chosen);
       return 0;
      }
